Split plotting files
