+++ src/qibocal/plots/__init__.py
@@ -1,3 +1,8 @@
 # -*- coding: utf-8 -*-
-from qibocal.plots.heatmaps import *
-from qibocal.plots.scatters import *
+from qibocal.plots.allXY import *
+from qibocal.plots.calibrate_qubit_states import *
+from qibocal.plots.flipping import *
+from qibocal.plots.rabi import *
+from qibocal.plots.ramsey import *
+from qibocal.plots.spectroscopies import *
+from qibocal.plots.t1 import *
