+++ src/qibocal/plots/ramsey.py
@@ -1,0 +1,115 @@
+# -*- coding: utf-8 -*-
+import numpy as np
+import plotly.graph_objects as go
+from plotly.subplots import make_subplots
+
+from qibocal.data import Data, DataUnits
+from qibocal.fitting.utils import ramsey
+
+
+# For Ramsey oscillations
+def time_msr(folder, routine, qubit, format):
+    try:
+        data = DataUnits.load_data(folder, routine, format, f"data_q{qubit}")
+    except:
+        data = DataUnits(
+            name=f"data_q{qubit}", quantities={"wait": "ns", "t_max": "ns"}
+        )
+    try:
+        data_fit = Data.load_data(folder, routine, format, f"fit_q{qubit}")
+    except:
+        data_fit = DataUnits()
+
+    fig = make_subplots(
+        rows=1,
+        cols=1,
+        horizontal_spacing=0.1,
+        vertical_spacing=0.1,
+        subplot_titles=("MSR (V)",),
+    )
+
+    fig.add_trace(
+        go.Scatter(
+            x=data.get_values("wait", "ns"),
+            y=data.get_values("MSR", "uV"),
+            name="Ramsey",
+        ),
+        row=1,
+        col=1,
+    )
+
+    # add fitting trace
+    if len(data) > 0 and len(data_fit) > 0:
+        timerange = np.linspace(
+            min(data.get_values("wait", "ns")),
+            max(data.get_values("wait", "ns")),
+            2 * len(data),
+        )
+        params = [i for i in list(data_fit.df.keys()) if "popt" not in i]
+        fig.add_trace(
+            go.Scatter(
+                x=timerange,
+                y=ramsey(
+                    timerange,
+                    data_fit.get_values("popt0"),
+                    data_fit.get_values("popt1"),
+                    data_fit.get_values("popt2"),
+                    data_fit.get_values("popt3"),
+                    data_fit.get_values("popt4"),
+                ),
+                name="Fit",
+                line=go.scatter.Line(dash="dot"),
+            ),
+            row=1,
+            col=1,
+        )
+
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.30,
+                showarrow=False,
+                text=f"Estimated {params[1]} is {data_fit.df[params[1]][0]:.3f} Hz.",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.20,
+                showarrow=False,
+                text=f"Estimated {params[0]} is {data_fit.df[params[0]][0]:.1f} ns",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+
+        fig.add_annotation(
+            dict(
+                font=dict(color="black", size=12),
+                x=0,
+                y=-0.25,
+                showarrow=False,
+                text=f"Estimated {params[2]} is {data_fit.df[params[2]][0]:.3f} Hz",
+                textangle=0,
+                xanchor="left",
+                xref="paper",
+                yref="paper",
+            )
+        )
+
+    fig.update_layout(
+        showlegend=True,
+        uirevision="0",  # ``uirevision`` allows zooming while live plotting
+        xaxis_title="Time (ns)",
+        yaxis_title="MSR (uV)",
+    )
+    return fig
